requirements.txt
# Core Speech-to-Text Dependencies
faster-whisper>=1.1.1
openai-whisper>=20250625

# Audio Processing
sounddevice>=0.5.2
soundfile>=0.13.1

# GUI Dependencies
# tkinter (included with Python)

# GPU Acceleration (Optional)
torch>=2.7.1
torchvision>=0.22.1
openvino>=2025.2.0

# Core Python Dependencies
numpy>=2.2.6
setuptools>=80.9.0
packaging>=25.0
requests>=2.32.4
tqdm>=4.67.1
PyYAML>=6.0.2

# Audio Format Support
av>=15.0.0

# Hugging Face Model Hub
huggingface-hub>=0.33.4

# Additional Processing
ctranslate2>=4.6.0
onnxruntime>=1.22.1

# Note: For GPU acceleration, install platform-specific packages:
# NVIDIA GPUs: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# AMD GPUs: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm5.7
# Intel GPUs: pip install openvino-dev